[CS-4015]: Migrate useAppState to .ts
