--- src/hooks/useAppState.js
@@ -1,27 +1,0 @@
-import { useEffect, useState } from 'react';
-import { AppState } from 'react-native';
-import usePrevious from './usePrevious';
-
-import { AppStateType } from '@cardstack/types';
-
-export default function useAppState() {
-  const [appState, setAppState] = useState(AppState.currentState);
-  const prevAppState = usePrevious(appState);
-
-  function onChange(newState) {
-    setAppState(newState);
-  }
-
-  useEffect(() => {
-    AppState.addEventListener('change', onChange);
-    return () => AppState.removeEventListener('change', onChange);
-  }, []);
-
-  return {
-    appState,
-    justBecameActive:
-      appState === AppStateType.active &&
-      prevAppState &&
-      prevAppState !== AppStateType.active,
-  };
-}
