+++ src/hooks/useClipboard.js
@@ -1,6 +1,6 @@
 import Clipboard from '@react-native-community/clipboard';
 import { useCallback, useEffect, useState } from 'react';
-import useAppState from './useAppState';
+import { useAppState } from './useAppState';
 import { deviceUtils } from '@rainbow-me/utils';
 
 const listeners = new Set();
